+++ pkg/index/index.go
@@ -441,4 +441,11 @@ type GetSearcher func(location databasev1.IndexRule_Type) (Searcher, error)
 type Filter interface {
 	fmt.Stringer
 	Execute(getSearcher GetSearcher, seriesID common.SeriesID, timeRange *RangeOpts) (posting.List, posting.List, error)
+	ShouldNotSkip(tagFamilyFilters FilterOp) bool
+}
+
+// FilterOp is an interface for filtering operations based on skipping index.
+type FilterOp interface {
+	Eq(tagName string, tagValue string) bool
+	Range(tagName string, rangeOpts RangeOpts) bool
 }